requirements.txt
pvporcupine>=3.0.0
pyaudio>=0.2.11
speechrecognition>=3.10.0
python-dotenv>=1.0.0
pygame>=2.5.0
numpy>=1.21.0
rpi-ws281x>=5.0.0